pyproject.toml
[project]
name = "face-sorter"
version = "0.1.0"
description = "A Python face recognition and sorting application using InsightFace, FAISS, and HDBSCAN"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Face Sorter Contributors" }
]
keywords = ["face-recognition", "computer-vision", "clustering", "image-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "insightface>=0.7.3",
    "faiss-cpu>=1.7.4",
    "scikit-learn>=1.0",
    "pymongo>=4.0",
    "opencv-python>=4.5",
    "Pillow>=9.0",
    "pillow-heif>=0.10",
    "torch>=1.10",
    "transformers>=4.20",
    "numpy>=1.20",
    "pydantic>=2.0",
    "python-dotenv>=0.19",
    "pydantic-settings>=2.0",
    "tqdm>=4.60",
    "psutil>=5.8",
    "onnxruntime>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "line-profiler>=3.5",
    "black>=22.0",
    "mypy>=0.950",
    "ruff>=0.1.0",
]

[project.scripts]
face-sorter = "face_sorter.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "line-profiler>=3.5",
    "black>=22.0",
    "mypy>=0.950",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
